Add Card object to all GameObjects with their own value.
