--- FightGame/Assets/Scripts/Card10Value.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Card10Value : MonoBehaviour {
-    public int value;
-
-	// Use this for initialization
-	void Start () {
-        value = 10;
-	}
-}
