--- FightGame/Assets/Scripts/Card10Value.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5ca22c8d3c83d435d8d391fee04ab1d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
