--- FightGame/Assets/Scripts/CardJValue.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CardJValue : MonoBehaviour {
-    public int value;
- 	// Use this for initialization
-	void Start () {
-        value = 11;
-	}
-}
