--- FightGame/Assets/Scripts/CardJValue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 04c7c593650ed48c3a059ac4566b4928
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
